Convert the mimic-ring stick quantizer through a signed integer
ftCo_800B0918 quantizes the leader's sticks and c-stick into Nana's
mimic ring as u8 bytes. MWCC does the conversion with fctiwz (signed,
wrapping) and stores the low byte, but the hosted inlineM0 returned the
float directly into u8: converting a negative float to an unsigned type
is undefined C, and while x86 happens to wrap like retail, arm64 and
wasm saturate to zero. Every leftward or downward sample silently
vanished from the ring on those hosts - in the live viewer Nana would
not mirror wavedashes (the down-diagonal died) and ignored c-stick
smashes entirely, while the x86 validation container never saw it.

Converting through s32 before the u8 store keeps the retail modulo
semantics on every host. Verified: the same batch-API scenario now
mirrors a c-stick downsmash and a wavedash six frames behind Popo on
Mac native and on the rebuilt wasm bundle; the container aggregate is
bit-unchanged (271 replays, 203 pass / 68 classified / 0 fail) and both
viewer smokes pass.

+++ src/melee/ft/chara/ftCommon/ftCo_0A01.c
@@ -7625,10 +7625,15 @@ static inline void ftCo_800B0760_dontinline(Fighter* fp)
 /// @todo Maybe a macro?
 static inline u8 inlineM0(float x)
 {
+    // MWCC converts through fctiwz (signed, wrapping) and stores the low
+    // byte. A direct float->u8 conversion of the negative half is undefined
+    // and saturates to zero on arm64 and wasm, silencing every leftward or
+    // downward sample in the follower's mimic ring; converting through s32
+    // keeps the retail modulo semantics on every host.
     if (x >= 0) {
-        return 127.0F * x;
+        return (u8) (s32) (127.0F * x);
     } else {
-        return 128.0F * x;
+        return (u8) (s32) (128.0F * x);
     }
 }
